refactor(shared): extract time utilities from runlog.py to shared/time_utils.py
Move utc_now_iso() and parse_iso8601() from adapters/persistence/runlog.py
to shared/time_utils.py. These are general-purpose time utilities used by
7+ modules across use_cases/, infra/, and adapters/ — embedding them in the
persistence adapter created false coupling.

Updated all 6 production importers and 3 test importers.

Closes #699
